make disabled test CE only
